DB-9537 Extract YARN profile to its own module
Add sparkHistoryServer script
