+++ pom.xml
@@ -890,6 +890,7 @@
                 <module>hbase_pipeline</module>
                 <module>hbase_sql</module>
                 <module>platform_it</module>
+                <module>standalone</module>
                 <module>spark_sql</module>
                 <module>splice_spark</module>
                 <module>scala_util</module>
@@ -926,6 +927,7 @@
                 <module>hbase_pipeline</module>
                 <module>hbase_sql</module>
                 <module>platform_it</module>
+                <module>standalone</module>
                 <module>spark_sql</module>
                 <module>splice_spark</module>
                 <module>scala_util</module>
@@ -962,6 +964,7 @@
                 <module>hbase_pipeline</module>
                 <module>hbase_sql</module>
                 <module>platform_it</module>
+                <module>standalone</module>
                 <module>spark_sql</module>
                 <module>splice_spark</module>
                 <module>scala_util</module>
